Added rest of the week 4 tasks

+++ controllers/blogs.js
@@ -1,59 +1,57 @@
 const blogsRouter = require('express').Router()
 const Blog = require('../models/blog')
-
-/*
-blogsRouter.get('/', (request, response) => {
-    Blog.find({}).then(blogs => {
-        response.json(blogs)
-    })
-})
-*/
+const middleware = require('../utils/middleware')
 
 blogsRouter.get('/', async (request, response) => {
-  const blogs = await Blog.find({})
+  const blogs = await Blog.find({}).populate('user', {username: 1, name: 1})
   response.json(blogs)
 })
 
-/*
-blogsRouter.post('/', (request, response, next) => {
-    const blog = new Blog(request.body)
-    blog
-      .save()
-      .then(result => {
-        response.status(201).json(result)
-      })    
-})
-*/
+blogsRouter.post('/', middleware.userExtractor, async (request, response) => {
+  const body = request.body  
+  const user = request.user
 
-blogsRouter.post('/', async (request, response, next) => {
-  const body = request.body
   const blog = new Blog({
     title: body.title,
     author: body.author,
     url: body.url,
-    likes: body.likes || 0 
+    likes: body.likes || 0,
+    user: user._id 
   })
+ 
+  const savedBlog = await blog.save()
+  user.blogs = user.blogs.concat(savedBlog._id)
+  await user.save()
 
-  try {
-    const savedBlog = await blog.save()
-    response.status(201).json(savedBlog)
-  } catch(exception) {
-    next(exception)
-  }
-
+  response.status(201).json(savedBlog)
+  
 })
 
-blogsRouter.delete('/:id', async (request, response) => {
-  await Blog.findByIdAndDelete(request.params.id)
-  response.status(204).end()
+blogsRouter.delete('/:id', middleware.userExtractor, async (request, response) => {
+  
+  // User validation is handled in userExtractor middleware
+  const user = request.user
+
+  const blog = await Blog.findById(request.params.id)
+  if (blog.user.toString() === user.id.toString()) {
+      
+    // Delete blog
+    await Blog.findByIdAndDelete(request.params.id)
+    
+    // Delete blog from users
+    user.blogs = user.blogs.filter(blog => blog.id !== request.params.id)
+    await user.save()
+    response.status(204).end()
+
+  } else {
+    return response.status(401).json({error: 'Unauthorized blog'})
+  }
+    
 })
 
 blogsRouter.put('/:id', async (request, response) => {
-   
+  
   const body = request.body
-  //console.log('Request body')
-  //console.log(body)
-
   const blog = {
     title: body.title,
     author: body.author,
@@ -61,15 +59,9 @@ blogsRouter.put('/:id', async (request, response) => {
     likes: body.likes
   }
 
-  const updatedBlog = await Blog.findByIdAndUpdate(request.params.id, blog, {new: true})
-  
-  //console.log('Updated blog:')
-  //console.log(updatedBlog)
-  
+  const updatedBlog = await Blog.findByIdAndUpdate(request.params.id, blog, {new: true})  
   response.status(204).json(updatedBlog)
 
 })
-
-
 
 module.exports = blogsRouter
